added better structure, setup tailwind with purge css, added navigation bar, removed unnecessary files
