+++ package.json
@@ -18,13 +18,10 @@
   "scripts": {
     "build:tailwind": "postcss src/tailwind.css -o src/tailwind.generated.css",
     "watch:tailwind": "postcss -w src/tailwind.css -o src/tailwind.generated.css",
-  
     "start": "run-p watch:tailwind start:react",
     "start:react": "react-scripts start",
-  
     "prebuild": "npm run build:tailwind",
     "build": "react-scripts build",
-  
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
