--- public/manifest.json
@@ -1,6 +1,0 @@
-{
-  "short_name": "Everplay",
-  "name": "Everplay mood list",
-  "start_url": ".",
-  "display": "standalone"
-}
