--- src/Components/Main/Main.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-class Main extends React.Component {
-  render(){
-    return <h1>This is the main component</h1>
-  }
-}
-
-export default Main;
