+++ src/components/Main.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import { Navigation } from './';
+
+class Main extends React.Component {
+  render(){
+    return(
+      <React.Fragment>
+        <Navigation />
+      </React.Fragment>
+    );
+  }
+}
+
+export default Main;
