+++ src/components/index.js
@@ -1,0 +1,3 @@
+export { default as Main } from './Main';
+export { default as Moodbar } from './Moodbar';
+export { default as Navigation } from './Navigation';
