+++ src/images/index.js
@@ -1,0 +1,2 @@
+export { default as Logo } from './Logo.svg';
+export { default as User} from './User.svg';
