+++ src/tailwind.css
@@ -1,3 +1,5 @@
-@tailwind base;
-@tailwind components;
+/* purgecss start ignore */
+@tailwind  base;
+@tailwind  components;
+/* purgecss end ignore */
 @tailwind utilities;
